+++ src/client/src/pages/detail/form/stratigraphy/lithology/lithologyLayers/descriptionLayers/descriptionDisplay.jsx
@@ -1,10 +1,8 @@
-import { useTranslation } from "react-i18next";
 import { Stack, Typography } from "@mui/material";
 import { formatNumberForDisplay } from "../../../../../../../components/form/formUtils.js";
 
 const DescriptionDisplay = props => {
   const { item, layerHeight } = props;
-  const { t, i18n } = useTranslation();
 
   const fontSize = layerHeight >= 5 ? 13 : (4 * layerHeight) / 1.1;
 
@@ -18,23 +16,16 @@ const DescriptionDisplay = props => {
         {formatNumberForDisplay(item.fromDepth)} m MD
       </Typography>
       {layerHeight >= 10 && (
-        <>
-          <Typography
-            sx={{
-              fontWeight: "bold",
-              overflow: "auto",
-              display: "-webkit-box",
-              WebkitLineClamp: "2",
-              WebkitBoxOrient: "vertical",
-            }}>
-            {item.description}
-          </Typography>
-          {item.id !== null && (
-            <Typography variant="subtitle2">
-              {t("description_quality")}: {item.descriptionQuality?.[i18n.language] ?? "-"}
-            </Typography>
-          )}
-        </>
+        <Typography
+          sx={{
+            fontWeight: "bold",
+            overflow: "hidden",
+            display: "-webkit-box",
+            WebkitLineClamp: "2",
+            WebkitBoxOrient: "vertical",
+          }}>
+          {item.description}
+        </Typography>
       )}
       <Typography
         variant="subtitle1"