+++ tests/api/Controllers/LithologicalDescriptionControllerTest.cs
@@ -70,8 +70,7 @@ public class LithologicalDescriptionControllerTest
         Assert.AreEqual(9_000_015, lithologicalDescription.Id);
         Assert.AreEqual(50, lithologicalDescription.FromDepth);
         Assert.AreEqual(60, lithologicalDescription.ToDepth);
-        Assert.AreEqual("Metrics Mountains Practical", lithologicalDescription.Description);
-        Assert.AreEqual(9000, lithologicalDescription.DescriptionQualityId);
+        Assert.AreEqual("Sleek Metrics Mountains", lithologicalDescription.Description);
         Assert.AreEqual(6_000_001, lithologicalDescription.StratigraphyId);
     }
 
@@ -88,8 +87,6 @@ public class LithologicalDescriptionControllerTest
             FromDepth = 90,
             ToDepth = 100,
             Id = 9_000_069,
-            DescriptionQuality = null,
-            DescriptionQualityId = 9001,
             Updated = new DateTime(2021, 3, 31, 16, 55, 02).ToUniversalTime(),
             UpdatedBy = null,
             UpdatedById = 5,
@@ -104,14 +101,13 @@ public class LithologicalDescriptionControllerTest
             Created = new DateTime(2021, 2, 14, 8, 55, 34).ToUniversalTime(),
             StratigraphyId = 6_000_010,
             Description = "Freddy ate more cucumber than Maria.",
-            DescriptionQualityId = 9003,
         };
 
         var lithologicalDescriptionToEdit = context.LithologicalDescriptions.Single(c => c.Id == id);
         Assert.AreEqual(1, lithologicalDescriptionToEdit.CreatedById);
-        Assert.AreEqual(5, lithologicalDescriptionToEdit.UpdatedById);
+        Assert.AreEqual(1, lithologicalDescriptionToEdit.UpdatedById);
         Assert.AreEqual(6_000_006, lithologicalDescriptionToEdit.StratigraphyId);
-        Assert.AreEqual("Libyan Dinar 1080p leading edge", lithologicalDescriptionToEdit.Description);
+        Assert.AreEqual("frame Libyan Dinar 1080p", lithologicalDescriptionToEdit.Description);
 
         // Update LithologicalDescription
         var response = await controller.EditAsync(newLithologicalDescription);
@@ -124,7 +120,6 @@ public class LithologicalDescriptionControllerTest
         Assert.AreEqual(1, updatedLithologicalDescription.UpdatedById);
         Assert.AreEqual(6_000_010, updatedLithologicalDescription.StratigraphyId);
         Assert.AreEqual("Freddy ate more cucumber than Maria.", updatedLithologicalDescription.Description);
-        Assert.AreEqual(9003, updatedLithologicalDescription.DescriptionQualityId);
     }
 
     [TestMethod]
@@ -159,7 +154,6 @@ public class LithologicalDescriptionControllerTest
             Created = new DateTime(2022, 11, 3, 14, 20, 09).ToUniversalTime(),
             StratigraphyId = 6_000_010,
             Description = "SPOLYP",
-            DescriptionQualityId = 9003,
         };
 
         var response = await controller.CreateAsync(lithologicalDescription);